// admin/src/app/core/api/model/userDTO.ts
/**
 * Univers des Âmes API
 * API REST pour la boutique et les services de l\'Univers des Âmes
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserDTO { 
    id?: number;
    email?: string;
    firstName?: string;
    lastName?: string;
    phone?: string;
    address?: string;
    city?: string;
    postalCode?: string;
    country?: string;
}